tests: fold gate findings — shared row asserts, one artifact pass, gate accessor
/simplify fold: assertRenders/assertFails/cleanSlate/fixture move to
RegressRowAsserts (compiled into both guard targets via HARNESS_SRCS) so
the two suites cannot drift; the three binary-framework rows share one
per-process artifact generation; the fail-on-gate-vs-skip-locally policy
gets its general name (RequiredGateEnforcement.enforced) instead of
riding the sim-specific flag, which now delegates to it.
