+++ previewsmcp/Tests/CLIIntegrationTests/BUILD.bazel
@@ -7,6 +7,7 @@ HARNESS_SRCS = [
     "CoreSimulatorHygiene.swift",
     "DaemonTestLock.swift",
     "PipeBuffer.swift",
+    "RegressRowAsserts.swift",
 ]
 
 swift_test(
